copy fixes to fresh_air_control

+++ fancoil_makeup_aux/include/modbus_client.h
@@ -4,7 +4,7 @@
 
 typedef uint8_t (*set_fancoil_func)(bool, uint8_t);
 
-void modbus_client_init(UCHAR slave_id, ULONG baud, uint8_t *iso_input_state,
+void modbus_client_init(UCHAR slave_id, ULONG baud, uint16_t *iso_input_state,
                         set_fancoil_func set_fancoil);
 
 void modbus_poll();
